fetch: return GraphQL data alongside path-level errors;
 review nits transport short-circuited on any `errors`, but the org/user
 metadata probe always returns NOT_FOUND for the unused namespace -- so
 every real fetch failed.

 Now return `data` when present, surface GraphQLError only when data is
 null.

 + testlib data + errors responder and transport/metadata regression.

 review nits:
   - fieldValues 30 -> 100; document 100-field cap on meta + items (no
     overflow)

   - project number (u64 -> u32) (GitHub Int is 32-bit)

   - reject [colors.values] set with empty [colors] github-name

+++ src/error/github.rs
@@ -72,7 +72,7 @@ pub enum GitHubError {
     #[error("project #{number} not found under {owner_kind} '{owner}'")]
     ProjectNotFound {
         owner: String,
-        number: u64,
+        number: u32,
         owner_kind: OwnerKind,
     },
